+++ cli/build.rs
@@ -68,6 +68,9 @@ fn create_compiler_snapshot(
     "deno.broadcast_channel",
     deno_broadcast_channel::get_declaration(),
   );
+  op_crate_libs.insert("deno.net", deno_net::get_declaration());
+  op_crate_libs
+    .insert("deno.net_unstable", deno_net::get_unstable_declaration());
 
   // ensure we invalidate the build properly.
   for (_, path) in op_crate_libs.iter() {
@@ -301,6 +304,14 @@ fn main() {
   println!(
     "cargo:rustc-env=DENO_BROADCAST_CHANNEL_LIB_PATH={}",
     deno_broadcast_channel::get_declaration().display()
+  );
+  println!(
+    "cargo:rustc-env=DENO_NET_LIB_PATH={}",
+    deno_net::get_declaration().display()
+  );
+  println!(
+    "cargo:rustc-env=DENO_NET_UNSTABLE_LIB_PATH={}",
+    deno_net::get_unstable_declaration().display()
   );
 
   println!("cargo:rustc-env=TARGET={}", env::var("TARGET").unwrap());
